Use the abstractions from `hound-javascript`
Which allows us to have a uniform API to:

- Send violations back to Hound.
- Report invalid configuration files back to Hound, to let the end user
  know.

https://trello.com/c/3BwQ9f4y
