+++ lib/config.js
@@ -1,6 +1,5 @@
 var configFile = require("eslint/lib/config/config-file");
 var fs = require("fs");
-var queue = require("./queue");
 var rules = require("eslint/lib/rules");
 var temp = require("temp").track();
 var util = require("eslint/lib/util");
