+++ lib/linter.js
@@ -1,10 +1,8 @@
 var EsLint = require("eslint").linter;
 var Config = require("./config");
-var ReportInvalidConfig = require("./report-invalid-config");
 
-function Linter(resultQueues) {
-  this.completedFileReviewQueue = resultQueues.complete;
-  this.reportInvalidConfigQueue = resultQueues.invalid;
+function Linter(houndJavascript) {
+  this.houndJavascript = houndJavascript;
 }
 
 Linter.prototype.lint = function(payload) {
@@ -20,7 +18,7 @@ Linter.prototype.lint = function(payload) {
       return { line: error.line, message: error.message };
     });
 
-    return this.completedFileReviewQueue.enqueue({
+    return this.houndJavascript.completeFileReview({
       violations: violations,
       filename: payload.filename,
       commit_sha: payload.commit_sha,
@@ -28,10 +26,11 @@ Linter.prototype.lint = function(payload) {
       patch: payload.patch,
     });
   } else {
-    var reportInvalidConfig = new ReportInvalidConfig(
-      this.reportInvalidConfigQueue
-    );
-    return reportInvalidConfig.run(payload);
+    return this.houndJavascript.reportInvalidConfig({
+      pull_request_number: payload.pull_request_number,
+      commit_sha: payload.commit_sha,
+      linter_name: "eslint",
+    });
   }
 };
