+++ lib/test-queue.js
@@ -1,4 +1,4 @@
-var Queue = require("./queue");
+var Queue = require("hound-javascript/lib/queue");
 
 module.exports = function(redis) {
   var inbound = new Queue({
