+++ tests/linter-test.js
@@ -1,9 +1,12 @@
+var Redis = require("fakeredis");
+var HoundJavascript = require("hound-javascript");
+var lastJob = require("./helpers/redis").lastJob;
+
 var Linter = require("../lib/linter");
-var Config = require("../lib/config");
 
 QUnit.module("Linter");
 
-test("ESLint linting", function() {
+asyncTest("ESLint linting", function() {
   var payload = {
     content: "var foo",
     config: "{ \"rules\": { \"semi\": 2 } }",
@@ -12,24 +15,35 @@ test("ESLint linting", function() {
     pull_request_number: "pull_request_number",
     patch: "patch",
   };
-  var linter = buildLinter();
+  var redis = Redis.createClient();
+  var houndJavascript = new HoundJavascript(redis);
+  var linter = new Linter(houndJavascript);
 
-  var job = linter.lint(payload);
-  var violation = job.violations[0];
+  linter.lint(payload).then(function() {
+    lastJob(redis, "high", function(job) {
+      start();
 
-  ok(violation.message.match(/semicolon/i), "includes the proper message");
-  equal(violation.line, 1, "includes the proper line");
-  equal(job.filename, payload.filename, "passes through filename");
-  equal(job.commit_sha, payload.commit_sha, "passes through commit_sha");
-  equal(
-    job.pull_request_number,
-    payload.pull_request_number,
-    "passes through pull_request_number"
-  );
-  equal(job.patch, payload.patch, "passes through patch");
+      equal(
+        job.class,
+        "CompletedFileReviewJob",
+        "pushes the proper job type"
+      );
+      deepEqual(
+        job.args[0],
+        {
+          violations: [ { line: 1, message: 'Missing semicolon.' } ],
+          filename: 'filename',
+          commit_sha: 'commit_sha',
+          pull_request_number: 'pull_request_number',
+          patch: 'patch',
+        },
+        "pushes a job onto the queue"
+      );
+    });
+  });
 });
 
-test("Reporting invalid configuration file", function() {
+asyncTest("Reporting invalid configuration file", function() {
   var payload = {
     content: "var foo",
     config: "---\nyaml: is good\ntrue/false/syntax/error",
@@ -38,42 +52,30 @@ test("Reporting invalid configuration file", function() {
     pull_request_number: "pull_request_number",
     patch: "patch",
   };
-  var linter = buildLinter();
+  var redis = Redis.createClient();
+  var houndJavascript = new HoundJavascript(redis);
+  var linter = new Linter(houndJavascript);
 
   var job = linter.lint(payload);
 
-  equal(
-    job.pull_request_number,
-    payload.pull_request_number,
-    "passes through pull_request_number"
-  );
-  equal(
-    job.commit_sha,
-    payload.commit_sha,
-    "passes through commit_sha"
-  );
-  equal(
-    job.linter_name,
-    "eslint",
-    "passes through linter_name"
-  );
-});
+  linter.lint(payload).then(function() {
+    lastJob(redis, "high", function(job) {
+      start();
 
-buildLinter = function() {
-  var completedFileReviewQueue = {
-    enqueue: function(job) {
-      return job;
-    },
-  };
-  var reportInvalidConfigQueue = {
-    enqueue: function(job) {
-      return job;
-    },
-  };
-  var linter = new Linter({
-    complete: completedFileReviewQueue,
-    invalid: reportInvalidConfigQueue,
+      equal(
+        job.class,
+        "ReportInvalidConfigJob",
+        "pushes the proper job type"
+      );
+      deepEqual(
+        job.args[0],
+        {
+          pull_request_number: "pull_request_number",
+          commit_sha: "commit_sha",
+          linter_name: "eslint",
+        },
+        "pushes a job onto the queue"
+      );
+    });
   });
-
-  return linter;
-};
+});
